+++ lib/sensu-mutator.rb
@@ -34,10 +34,10 @@ module Sensu
   class Mutator
     include Sensu::Plugin::Utils
     include Mixlib::CLI
-    option :map_v2_event_into_v1,
-           description: 'Enable 2.x to 1.4 event mapping. Alternatively set envvar SENSU_MAP_V2_EVENT_INTO_V1=1.',
+    option :map_go_event_into_ruby,
+           description: 'Enable Sensu Go to Sensu Ruby event mapping. Alternatively set envvar SENSU_MAP_GO_EVENT_INTO_RUBY=1.',
            boolean:     true,
-           long:        '--map-v2-event-into-v1'
+           long:        '--map-go-event-into-ruby'
 
     attr_accessor :argv
 
@@ -73,10 +73,10 @@ module Sensu
       mutator.read_event(STDIN)
 
       TRUTHY_VALUES = %w[1 t true yes y].freeze
-      automap = ENV['SENSU_MAP_V2_EVENT_INTO_V1'].to_s.downcase
+      automap = ENV['SENSU_MAP_GO_EVENT_INTO_RUBY'].to_s.downcase
 
-      if mutator.config[:map_v2_event_into_v1] || TRUTHY_VALUES.include?(automap)
-        new_event = mutator.map_v2_event_into_v1
+      if mutator.config[:map_go_event_into_ruby] || TRUTHY_VALUES.include?(automap)
+        new_event = mutator.map_go_event_into_ruby
         mutator.event = new_event
       end